requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.3.1
prompt-toolkit==3.0.39
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
six==1.16.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
